// sdk/src/generated/accounts/tickArrayState.ts
export type TickArrayState = {
  discriminator: ReadonlyUint8Array;
  poolId: Address;
  startTickIndex: number;
  ticks: Array<TickState>;
  initializedTickCount: number;
  recentEpoch: bigint;
  padding: ReadonlyUint8Array;
};